Added generic CSV file import feature that will work for any node.
Add the 'import' privilege to the getNodes() call in module.inc to
activate the import feature for a node. After granting this right to
a user, any type of CSV file can be imported.

Todo:
 - date support
 - validation of column values (it is now possible to store values that
   do not match the values of a listattribute)
 - validation of obligatory fields, duplicate fields etc.
 - Reporting to the user which lines have failed.

The importer uses initial_values() of the importing node, so default
values for fields not in the CSV file can be set.