+ Upgraded Prototype to 1.6.0.2, Scriptaculous to 1.8.1 and Window.js to   1.3.2. Might cause certain backwards incompatiblities! [Peter] - Removed Rico.js because it was almost never used and outdated. [Peter]
Because of the backwards incompatible changes please let me know if you find
something which is part of ATK that is now broken. Everything outside of
ATK should be fixed. Fixes are in most cases trivial and for 95% of the code
out there unneeded.

If you use Rico inside your application please install it as part of your
application. A new version exists (2.0 RC2) which should be compatible with
the latest Prototype version (doesn't use Prototype anymore).

At the moment the web20 theme is broken because it requires Rico. This theme
served as demo of ATK's frameless capabilities. Can this theme be removed
now we have the stillblue theme?