Stop passing relation destination filters in links
There is no way to do this securely.
For M2O->autocomplete in edit : the original record is retrieved in autocomplete
function and is used to parse m_destinationFilter.
Rename 'atkfilter' to 'atkforce' for 'add' links on relation (it presets value
in the destination form)
Still to remove : atkfilter in admin/export links (not tested yet), atkfilter in
TreeNode (but for TreeNode, I wait for integration of changes from 9.0).